fix: declared to which packages a few orphan files were.
These are where the ports to be used throughout the apps should be declared.

In Hex Arch you declare ports (which are basically interfaces for intended behaviour) and design adapters (things that implement said behaviour and plug to the ports) so that the code is not dependent on any one thing.

If done properly, this enables us to implement SQL and Redis for example with minimal overhead since coupling is minimized.