feat(models): Finall-friggin-ly.
Ok models are done. I made few changes as I went one of which is not using SIN as a primary ID (cause that's begging for problems IRL) and replacing that by a regular int id.

More thought could have been given for the ID scheme, but frankly couldn't care less. This works good enough.

At this point the models are done as far as a I am aware, so it should be fixes/corrections from now on.