feat(employee.go): Exported the fields of the Employee field.
Similarly to how'd you do it in Rust, I started by defining all the attributes of the fields as prive (lower case name).

Now as a need to basis, I will export those fields. Likely a lot of the stuff that needs to be stored in a database will have to be exported anyways, but now things that are not needed will remain private.